renamed filter to view
