+++ src/components/VerticalGeneSetScoresView.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 
 import DrawFunctions from '../functions/DrawFunctions';
 import { VERTICAL_GENESET_DETAIL_WIDTH ,VERTICAL_GENESET_SUPPRESS_WIDTH } from '../components/XenaGeneSetApp';
-import {FILTER_ENUM} from './FilterSelector';
+import {FILTER_ENUM} from './ViewSelector';
 import CanvasDrawing from './CanvasDrawing';
 import {createAssociatedDataKey, findAssociatedData, findPruneData} from '../functions/DataFunctions';
 import {clusterSampleSort, selectedSampleGeneExpressionActivitySort} from '../functions/SortFunctions';
