+++ src/functions/DataFunctions.js
@@ -11,7 +11,7 @@ import { getGenesForNamedPathways } from './CohortFunctions';
 import {
   clusterSort, diffSort, scoreColumns, synchronizedSort, geneExpressionSort,
 } from './SortFunctions';
-import {FILTER_ENUM} from '../components/FilterSelector';
+import {FILTER_ENUM} from '../components/ViewSelector';
 
 export const DEFAULT_AMPLIFICATION_THRESHOLD = 2;
 export const DEFAULT_DELETION_THRESHOLD = -2;
