+++ tests/function/CohortFunctions.test.js
@@ -13,7 +13,7 @@ import TestDefaultCohortData from '../data/TestDefaultCohortData';
 import SamplePathway1 from '../data/SamplePathway1';
 import {UNASSIGNED_SUBTYPE} from '../../src/components/SubCohortSelector';
 import {calculateSelectedSubCohortSamples} from '../../src/functions/FetchFunctions';
-import {FILTER_ENUM} from '../../src/components/FilterSelector';
+import {FILTER_ENUM} from '../../src/components/ViewSelector';
 const TEST_MESOTHELIOMA_SAMPLES = ['TCGA-3H-AB3K-01','TCGA-3H-AB3L-01','TCGA-3H-AB3M-01','TCGA-3H-AB3O-01','TCGA-3H-AB3S-01','TCGA-3H-AB3T-01','TCGA-3H-AB3U-01','TCGA-3H-AB3X-01','TCGA-3U-A98D-01','TCGA-3U-A98E-01','TCGA-3U-A98F-01','TCGA-3U-A98G-01','TCGA-3U-A98H-01','TCGA-3U-A98I-01','TCGA-3U-A98J-01','TCGA-LK-A4NW-01','TCGA-LK-A4NY-01','TCGA-LK-A4NZ-01','TCGA-LK-A4O0-01','TCGA-LK-A4O2-01','TCGA-LK-A4O4-01','TCGA-LK-A4O5-01','TCGA-LK-A4O6-01','TCGA-LK-A4O7-01','TCGA-MQ-A4KX-01','TCGA-MQ-A4LC-01','TCGA-MQ-A4LI-01','TCGA-MQ-A4LJ-01','TCGA-MQ-A4LM-01','TCGA-MQ-A4LP-01','TCGA-MQ-A4LV-01','TCGA-MQ-A6BL-01','TCGA-MQ-A6BQ-01','TCGA-MQ-A6BR-01','TCGA-MQ-A6BS-01','TCGA-NQ-A57I-01','TCGA-NQ-A638-01','TCGA-SC-A6LM-01','TCGA-SC-A6LN-01','TCGA-SC-A6LP-01','TCGA-SC-A6LQ-01','TCGA-SC-A6LR-01','TCGA-SC-AA5Z-01','TCGA-SH-A7BC-01','TCGA-SH-A7BD-01','TCGA-SH-A7BH-01','TCGA-SH-A9CT-01','TCGA-SH-A9CU-01','TCGA-TS-A7OU-01','TCGA-TS-A7OY-01','TCGA-TS-A7OZ-01','TCGA-TS-A7P0-01','TCGA-TS-A7P1-01','TCGA-TS-A7P3-01','TCGA-TS-A7P6-01','TCGA-TS-A7P8-01','TCGA-TS-A7PB-01','TCGA-TS-A8AI-01','TCGA-TS-A8AS-01','TCGA-TS-A8AV-01','TCGA-TS-A8AY-01','TCGA-UD-AABZ-01','TCGA-UD-AAC1-01','TCGA-UD-AAC4-01','TCGA-UD-AAC5-01','TCGA-UD-AAC6-01','TCGA-UD-AAC7-01','TCGA-UT-A88C-01','TCGA-UT-A88D-01','TCGA-UT-A88E-01','TCGA-UT-A88G-01','TCGA-UT-A97Y-01','TCGA-XT-AASU-01','TCGA-YS-A95B-01','TCGA-YS-A95C-01','TCGA-YS-AA4M-01','TCGA-ZN-A9VO-01','TCGA-ZN-A9VP-01','TCGA-ZN-A9VQ-01','TCGA-ZN-A9VS-01','TCGA-ZN-A9VU-01','TCGA-ZN-A9VV-01','TCGA-ZN-A9VW-01'];
 const TEST_AVAILABLE_OVARIAN_SAMPLES = ['TCGA-04-1353-01','TCGA-04-1516-01','TCGA-04-1638-01','TCGA-04-1644-01','TCGA-04-1646-01','TCGA-04-1648-01','TCGA-04-1649-01','TCGA-04-1651-01','TCGA-04-1652-01','TCGA-04-1655-01','TCGA-09-1664-01','TCGA-09-1670-01','TCGA-09-1672-01','TCGA-09-1673-01','TCGA-09-1674-01','TCGA-09-1675-01','TCGA-13-2057-01','TCGA-13-2059-01','TCGA-13-2061-01','TCGA-13-2065-01','TCGA-13-2071-01','TCGA-20-1682-01','TCGA-20-1683-01','TCGA-20-1684-01','TCGA-20-1685-01','TCGA-20-1686-01','TCGA-20-1687-01','TCGA-23-1029-01','TCGA-23-1111-01','TCGA-23-1114-01','TCGA-23-1809-01','TCGA-23-2641-01','TCGA-23-2643-01','TCGA-23-2645-01','TCGA-24-1842-01','TCGA-24-1843-01','TCGA-24-1844-01','TCGA-24-1845-01','TCGA-24-1846-01','TCGA-24-1847-01','TCGA-24-1849-01','TCGA-24-1850-01','TCGA-29-1688-01','TCGA-29-1690-01','TCGA-29-1691-01','TCGA-29-1693-01','TCGA-29-1694-01','TCGA-29-1695-01','TCGA-29-1696-01','TCGA-29-1697-01','TCGA-29-1698-01','TCGA-29-1699-01','TCGA-29-1701-01','TCGA-29-1702-01','TCGA-29-1703-01','TCGA-29-1705-01','TCGA-29-1707-01','TCGA-29-1710-01','TCGA-29-1711-01','TCGA-29-1761-01','TCGA-29-1762-01','TCGA-29-1763-01','TCGA-29-1764-01','TCGA-29-1766-01','TCGA-29-1768-01','TCGA-29-1769-01','TCGA-29-1770-01','TCGA-29-1771-01','TCGA-29-1774-01','TCGA-29-1775-01','TCGA-29-1776-01','TCGA-29-1777-01','TCGA-29-1778-01','TCGA-29-1781-01','TCGA-29-1783-01','TCGA-29-1784-01','TCGA-29-1785-01','TCGA-29-2429-01','TCGA-29-2431-01','TCGA-29-2432-01','TCGA-29-2434-01','TCGA-29-2436-01','TCGA-30-1714-01','TCGA-30-1718-01','TCGA-30-1855-01','TCGA-30-1856-01','TCGA-30-1857-01','TCGA-36-2530-01','TCGA-36-2532-01','TCGA-36-2533-01','TCGA-36-2534-01','TCGA-36-2537-01','TCGA-36-2538-01','TCGA-36-2539-01','TCGA-36-2540-01','TCGA-36-2542-01','TCGA-36-2543-01','TCGA-36-2544-01','TCGA-36-2545-01','TCGA-36-2547-01','TCGA-36-2548-01','TCGA-36-2551-01','TCGA-36-2552-01','TCGA-42-2582-01','TCGA-42-2587-01','TCGA-42-2588-01','TCGA-42-2589-01','TCGA-42-2590-01','TCGA-42-2591-01','TCGA-57-1586-01','TCGA-59-2372-01','TCGA-61-1722-01','TCGA-61-1725-01','TCGA-61-1727-01','TCGA-61-1730-01','TCGA-61-1733-01','TCGA-61-1737-01','TCGA-61-1738-01','TCGA-61-1740-01','TCGA-61-1741-01','TCGA-61-1895-01','TCGA-61-1899-01','TCGA-61-1900-01','TCGA-61-1901-01','TCGA-61-1903-01','TCGA-61-1904-01','TCGA-61-1906-01','TCGA-61-1907-01','TCGA-61-1910-01','TCGA-61-1911-01','TCGA-61-1913-01','TCGA-61-1914-01','TCGA-61-1915-01','TCGA-61-2612-01','TCGA-61-2613-01','TCGA-61-2614-01'];
 const BREAST_SUBCOHORTS = {'name':'TCGA Breast Cancer (BRCA)','mutationDataSetId':'TCGA.BRCA.sampleMap/mutation_curated_wustl','copyNumberDataSetId':'TCGA.BRCA.sampleMap/Gistic2_CopyNumber_Gistic2_all_thresholded.by_genes','genomeBackgroundCopyNumber':{'host':'https://xenago.xenahubs.net','dataset':'cnv_sampleEvent','feature_event_K':'event_K','feature_total_pop_N':'total_pop_N'},'genomeBackgroundMutation':{'host':'https://xenago.xenahubs.net','dataset':'mutation_sampleEvent','feature_event_K':'event_K','feature_total_pop_N':'total_pop_N'},'amplificationThreshold':2,'deletionThreshold':-2,'host':'https://tcga.xenahubs.net','subCohorts':['BRCA.Basal','BRCA.Her2','BRCA.LumA','BRCA.LumB','BRCA.Normal'],'selectedSubCohorts':['BRCA.LumB']};
