+++ tests/function/FetchFunctions.test.js
@@ -1,7 +1,7 @@
 import expect from 'expect';
 import { getSamplesForCohort} from '../../src/functions/FetchFunctions';
 import {getCohortDetails} from '../../src/functions/CohortFunctions';
-import {FILTER_ENUM} from '../../src/components/FilterSelector';
+import {FILTER_ENUM} from '../../src/components/ViewSelector';
 
 function logError(err) {
   // eslint-disable-next-line no-console
